Refactor triplicate CLI config application code
