+++ packages/ladle/lib/cli/build.js
@@ -4,11 +4,11 @@ import path from "path";
 import { promises as fs } from "fs";
 import { globby } from "globby";
 import viteProd from "./vite-prod.js";
-import loadConfig from "./load-config.js";
 import debug from "./debug.js";
 import { getMetaJsonString } from "./vite-plugin/generate/get-meta-json.js";
 import { getEntryData } from "./vite-plugin/parse/get-entry-data.js";
 import getFolderSize from "./get-folder-size.js";
+import applyCLIConfig from "./apply-cli-config.js";
 
 /**
  * @param params {import("../shared/types").BuildParams}
@@ -16,25 +16,7 @@ import getFolderSize from "./get-folder-size.js";
 const build = async (params = {}) => {
   const startTime = performance.now();
   debug("Starting build command");
-  debug(`CLI theme: ${params.theme}`);
-  debug(`CLI stories: ${params.stories}`);
-  debug(`CLI out: ${params.outDir ? params.outDir : "undefined"}`);
-  params.config = params.config || ".ladle";
-  const configFolder = path.isAbsolute(params.config)
-    ? params.config
-    : path.join(process.cwd(), params.config);
-  const config = await loadConfig(configFolder);
-
-  // CLI flags override default and custom config files
-  config.addons.theme.defaultState = params.theme
-    ? params.theme
-    : config.addons.theme.defaultState;
-  config.stories = params.stories ? params.stories : config.stories;
-  config.viteConfig = params.viteConfig ? params.viteConfig : config.viteConfig;
-  config.outDir = params.outDir ? params.outDir : config.outDir;
-  debug(`Final config:\n${JSON.stringify(config, null, "  ")}`);
-  process.env["VITE_PUBLIC_LADLE_THEME"] = config.addons.theme.defaultState;
-  process.env["VITE_PUBLIC_STORIES"] = config.stories;
+  const { configFolder, config } = await applyCLIConfig(params);
   await viteProd(config, configFolder);
   const entryData = await getEntryData(await globby([config.stories]));
   const jsonContent = getMetaJsonString(entryData);
