+++ packages/ladle/lib/cli/cli.js
@@ -18,10 +18,7 @@ program
   .option("-p, --port [number]", "port to serve the application", strToInt)
   .option("--stories [string]", "glob to find stories")
   .option("--theme [string]", "theme light, dark or auto")
-  .option(
-    "--config [string]",
-    "folder where Ladle configs are located, default .ladle",
-  )
+  .option("--config [string]", "folder where config is located, default .ladle")
   .option("--viteConfig [string]", "file with Vite configuration")
   .action(async (params) => {
     await serve({ ...params, serve: params });
