+++ packages/ladle/lib/cli/serve.js
@@ -1,33 +1,15 @@
 #!/usr/bin/env node
 
-import path from "path";
 import viteDev from "./vite-dev.js";
-import loadConfig from "./load-config.js";
 import debug from "./debug.js";
+import applyCLIConfig from "./apply-cli-config.js";
 
 /**
  * @param params {import("../shared/types").DevParams}
  */
 const serve = async (params = {}) => {
   debug("Starting serve command");
-  debug(`CLI theme: ${params.theme}`);
-  debug(`CLI stories: ${params.stories}`);
-  debug(`CLI port: ${params.port ? params.port : "undefined"}`);
-  params.config = params.config || ".ladle";
-  const configFolder = path.isAbsolute(params.config)
-    ? params.config
-    : path.join(process.cwd(), params.config);
-  const config = await loadConfig(configFolder);
-  // CLI flags override default and custom config files
-  config.addons.theme.defaultState = params.theme
-    ? params.theme
-    : config.addons.theme.defaultState;
-  config.stories = params.stories ? params.stories : config.stories;
-  config.port = params.port ? params.port : config.port;
-  config.viteConfig = params.viteConfig ? params.viteConfig : config.viteConfig;
-  debug(`Final config:\n${JSON.stringify(config, null, "  ")}`);
-  process.env["VITE_PUBLIC_LADLE_THEME"] = config.addons.theme.defaultState;
-  process.env["VITE_PUBLIC_STORIES"] = config.stories;
+  const { configFolder, config } = await applyCLIConfig(params);
   await viteDev(config, configFolder);
 };
